Refactor config code into its own module and rename Config to Env
