+++ src/config/config.rs
@@ -1,0 +1,84 @@
+/* Pi-hole: A black hole for Internet advertisements
+*  (c) 2018 Pi-hole, LLC (https://pi-hole.net)
+*  Network-wide ad blocking via your own hardware.
+*
+*  API
+*  Config File Structure
+*
+*  This file is copyright under the latest version of the EUPL.
+*  Please see LICENSE file for your rights under this license. */
+
+use config::PiholeFile;
+
+/// The API config options
+#[derive(Deserialize, Default)]
+pub struct Config {
+    #[serde(default)]
+    file_locations: Files
+}
+
+impl Config {
+    /// Get the configured location of a file
+    pub fn file_location(&self, file: PiholeFile) -> &str {
+        match file {
+            PiholeFile::DnsmasqMainConfig => &self.file_locations.dnsmasq_main_config,
+            PiholeFile::Whitelist => &self.file_locations.whitelist,
+            PiholeFile::Blacklist => &self.file_locations.blacklist,
+            PiholeFile::Regexlist => &self.file_locations.regexlist,
+            PiholeFile::SetupVars => &self.file_locations.setup_vars,
+            PiholeFile::LocalVersions => &self.file_locations.local_versions,
+            PiholeFile::LocalBranches => &self.file_locations.local_branches
+        }
+    }
+}
+
+/// Defines the deserialization of the "file_locations" section of the config file.
+/// The default functions are generated by `default!`.
+#[derive(Deserialize)]
+pub struct Files {
+    #[serde(default = "default_dnsmasq_main_config")]
+    dnsmasq_main_config: String,
+    #[serde(default = "default_whitelist")]
+    whitelist: String,
+    #[serde(default = "default_blacklist")]
+    blacklist: String,
+    #[serde(default = "default_regexlist")]
+    regexlist: String,
+    #[serde(default = "default_setup_vars")]
+    setup_vars: String,
+    #[serde(default = "default_local_versions")]
+    local_versions: String,
+    #[serde(default = "default_local_branches")]
+    local_branches: String
+}
+
+impl Default for Files {
+    fn default() -> Self {
+        Files {
+            dnsmasq_main_config: default_dnsmasq_main_config(),
+            whitelist: default_whitelist(),
+            blacklist: default_blacklist(),
+            regexlist: default_regexlist(),
+            setup_vars: default_setup_vars(),
+            local_versions: default_local_versions(),
+            local_branches: default_local_branches()
+        }
+    }
+}
+
+/// Create an `fn() -> String` default function for deserialization
+macro_rules! default {
+    ($fn_name:ident, $variant:ident) => {
+        fn $fn_name() -> String {
+            PiholeFile::$variant.default_location().to_owned()
+        }
+    };
+}
+
+default!(default_dnsmasq_main_config, DnsmasqMainConfig);
+default!(default_whitelist, Whitelist);
+default!(default_blacklist, Blacklist);
+default!(default_regexlist, Regexlist);
+default!(default_setup_vars, SetupVars);
+default!(default_local_versions, LocalVersions);
+default!(default_local_branches, LocalBranches);
